Fix tests

+++ auth/api_token_test.go
@@ -17,13 +17,13 @@ func (s *S) TestAPIAuth(c *check.C) {
 	c.Assert(t.Token, check.Equals, APIKey)
 	c.Assert(t.UserEmail, check.Equals, user.Email)
 
-	c.Assert(user.APIKeyUsageCounter, check.Equals, 0)
+	c.Assert(user.APIKeyUsageCounter, check.Equals, int64(0))
 
 	err = user.Reload()
 	c.Assert(err, check.IsNil)
 
 	c.Assert(user.APIKeyLastAccess.IsZero(), check.Equals, false)
-	c.Assert(user.APIKeyUsageCounter, check.Equals, 1)
+	c.Assert(user.APIKeyUsageCounter, check.Equals, int64(1))
 
 }
